FRO-235: Refactor- Break down question card into smaller components for better maintainability
